Complete refactor of transactions
If we think about transactions as IReadWrites that are just wrappers
around their underlying database, then it makes the method signatures
for everything much simpler - as well as the implementation.

To do a transaction, you now just call db.BeginTransaction(), which gives
you a new object (the transaction object) that also implements IReadWrite -
then you simply do all your CRUD operations directly on the transaction.
Because we have the transaction self-contained now, we can eliminate the
transaction option to the original database methods, and they no longer
have to be coded to take transactions into account.

Also, because having the ability to do transactions is seperate from
reading and writing, a new interface - ITransactable - was created that
derives from IReadWrite