Taxonomical additions
There are now TypedAnalyzers, and modifiable dbs can now utilize
a TypedAnalyzer as a Taxonomy - this will tell the db what structure
to use when creating new tables.

SqlModifier has been upgrades to act as a TypedAnalyzer.

There's a new TypedAnalyzer based on objects, allowing you to easily
create DB tables that follow the structure of objects